[issue-291] Fixed a lot of unittests
